# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ระบบ Blog + Auth + OTP + Like/Comment ที่เขียนด้วย Go + Gin",
        "title": "Blog API Platform",
        "contact": {
            "name": "Filly",
            "email": "fillybodyknow@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "เข้าสู่ระบบด้วย Username/Email และ Password",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authication"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"blog or blogg@example.com\"",
                        "description": "Username/Email*",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Strong@Password123\"",
                        "description": "Password*",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/register": {
            "post": {
                "description": "สร้างผู้ใช้งานใหม่ด้วย Username, Email และ Password",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authication"
                ],
                "summary": "สมัครสมาชิก",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"blogg@example.com\"",
                        "description": "Email*",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"blog\"",
                        "description": "Username*",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Strong@Password123\"",
                        "description": "Password*",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/verify-otp": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ยืนยัน OTP ของผู้ใช้",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authication"
                ],
                "summary": "ยืนยัน OTP",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"123456\"",
                        "description": "OTP*",
                        "name": "otp",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/auth/verify_otp": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ส่ง OTP ไปยัง Email ของผู้ใช้",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authication"
                ],
                "summary": "ส่ง OTP",
                "responses": {}
            }
        },
        "/posts": {
            "get": {
                "description": "แสดง Post ตาม Tags ไม่ต้องยืนยันตัวตน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "แสดง Post ตาม Tags",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"tag1,tag2,tag3\"",
                        "description": "Tags",
                        "name": "tags",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/posts/": {
            "get": {
                "description": "แสดง Post ทั้งหมด ไม่ต้องยืนยันตัวตน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "แสดง Post ทั้งหมด",
                "responses": {}
            }
        },
        "/posts/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ต้องยืนยันตัวตนก่อนจึงจะสามารถสร้าง Post ได้",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "สร้าง Post",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Hello World\"",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Hello World Content\"",
                        "description": "Content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"tag1,tag2,tag3\"",
                        "description": "Tags",
                        "name": "tags",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/posts/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "แสดง Post ของตัวเอง ไม่ต้องยืนยันตัวตน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "แสดง Post ของตัวเอง",
                "responses": {}
            }
        },
        "/posts/{post_id}": {
            "get": {
                "description": "แสดง Post ตาม ID ไม่ต้องยืนยันตัวตน",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "แสดง Post ตาม ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "แก้ไข Post ของตัวเอง ต้องยืนยันตัวตนก่อนจึงจะสามารถแก้ไข Post ได้",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "แก้ไข Post ของตัวเอง",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Hello World\"",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Hello World\"",
                        "description": "Content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"tag1,tag2,tag3\"",
                        "description": "Tags",
                        "name": "tags",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ลบ Post ของตัวเอง ต้องยืนยันตัวตนก่อนจึงจะสามารถลบ Post ได้",
                "tags": [
                    "Post"
                ],
                "summary": "ลบ Post ของตัวเอง",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/posts/{post_id}/comment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้าง Comment ด้วยการกรอก Content",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Create Comment",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"123456789012345678901234\"",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Hello World\"",
                        "description": "Content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts/{post_id}/comment/{comment_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "แก้ไข Comment ด้วนการกรอก Content",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Edit Comment",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"123456789012345678901234\"",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"123456789012345678901234\"",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Hello World\"",
                        "description": "Content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ลบ Comment ของตัวเอง",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Delete Comment",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"123456789012345678901234\"",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"123456789012345678901234\"",
                        "description": "Comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/posts/{post_id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Like โพสต์",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Like"
                ],
                "summary": "Like",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "UnLike โพสต์",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Like"
                ],
                "summary": "UnLike",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}