# internal/handler/authication.go
package handler

import (
	"time"

	"github.com/Fillybodyknow/blog-api/internal/models"
	"github.com/Fillybodyknow/blog-api/internal/service"
	"github.com/gin-gonic/gin"
	"go.mongodb.org/mongo-driver/bson/primitive"
)

type AuthHandler struct {
	AuthServiceInterface service.AuthServiceInterface
}

func NewAuthHandler(authService service.AuthServiceInterface) *AuthHandler {
	return &AuthHandler{AuthServiceInterface: authService}
}

// RegisterUser godoc
// @Summary สมัครสมาชิก
// @Description สร้างผู้ใช้งานใหม่ด้วย Username, Email และ Password
// @Tags Authication
// @Accept application/x-www-form-urlencoded
// @Produce json
// @Param email formData string true "Email*" example("blogg@example.com")
// @Param username formData string true "Username*" example("blog")
// @Param password formData string true "Password*" example("Strong@Password123")
// @Router /auth/register [post]
func (h *AuthHandler) RegisterUser(c *gin.Context) {

	var input models.RegisterInput

	if err := c.ShouldBind(&input); err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	user := models.User{
		Username:     input.Username,
		Email:        input.Email,
		PasswordHash: input.Password,
		Role:         "user",
		IsVerified:   false,
		CreatedAt:    time.Now(),
	}

	err := h.AuthServiceInterface.Register(c, &user)
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}
	c.JSON(200, gin.H{"message": "Registered successfully"})

}

// LoginUser godoc
// @Summary เข้าสู่ระบบ
// @Description เข้าสู่ระบบด้วย Username/Email และ Password
// @Tags Authication
// @Accept application/x-www-form-urlencoded
// @Produce json
// @Param username formData string true "Username/Email*" example("blog or blogg@example.com")
// @Param password formData string true "Password*" example("Strong@Password123")
// @Router /auth/login [post]
func (h *AuthHandler) LoginUser(c *gin.Context) {
	var input models.LoginUserInput

	if err := c.ShouldBind(&input); err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	var user models.User

	user, token, err := h.AuthServiceInterface.Login(c, input.Username, input.Password)
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	c.JSON(200, gin.H{"user_id": user.ID, "role": user.Role, "token": token})
}

// SendOTP godoc
// @Summary ส่ง OTP
// @Description ส่ง OTP ไปยัง Email ของผู้ใช้
// @Tags Authication
// @Security BearerAuth
// @Accept json
// @Produce json
// @Router /auth/verify_otp [get]
func (h *AuthHandler) OTP(c *gin.Context) {
	UserIDStr, _ := c.Get("user_id")
	objID, err := primitive.ObjectIDFromHex(UserIDStr.(string))
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}
	err = h.AuthServiceInterface.SendOTP(objID, c)
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	c.JSON(200, gin.H{"message": "ส่ง OTP สําเร็จ"})
}

// VerifyOTP godoc
// @Summary ยืนยัน OTP
// @Description ยืนยัน OTP ของผู้ใช้
// @Tags Authication
// @Security BearerAuth
// @Accept application/x-www-form-urlencoded
// @Produce json
// @Param otp formData string true "OTP*" example("123456")
// @Router /auth/verify-otp [post]
func (h *AuthHandler) VerifyOTP(c *gin.Context) {
	UserIDStr, _ := c.Get("user_id")
	objID, err := primitive.ObjectIDFromHex(UserIDStr.(string))
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}

	var input models.VerifyOTPInput
	if err := c.ShouldBind(&input); err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}
	err = h.AuthServiceInterface.VerifyOTP(objID, input.OTP)
	if err != nil {
		c.JSON(400, gin.H{"error": err.Error()})
		return
	}
	c.JSON(200, gin.H{"message": "ยืนยัน OTP สําเร็จ"})
}